Code Refactor: Separate files for snake env and configs
